[refactor/backend] CB-14: Altera ID com tipo Integer para Long

+++ backend/src/main/java/com/optimiza/clickbarber/model/Barbearia.java
@@ -6,7 +6,6 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.util.List;
-import java.util.Objects;
 
 @Table(name = "barbearias")
 @Entity
@@ -21,7 +20,7 @@ public class Barbearia {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String cnpj;
     private String nome;
