+++ backend/src/main/resources/db/migration/V3__cria_tabela_servicos.sql
@@ -3,7 +3,7 @@ CREATE TABLE servicos (
     nome VARCHAR(100) NOT NULL,
     preco NUMERIC(10, 2) NOT NULL,
     tempo_duracao_em_minutos INTEGER NOT NULL,
-    barbearia_id INTEGER NOT NULL,
+    barbearia_id BIGINT NOT NULL,
     ativo BOOLEAN NOT NULL DEFAULT true,
     CONSTRAINT fk_barbearia FOREIGN KEY (barbearia_id) REFERENCES barbearias(id)
 );
